version_7_programmatic_tool_execution/pyproject.toml
[project]
name = "version-7-programmatic-tool-execution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.6.5",
    "click>=8.3.1",
    "fastapi>=0.124.0",
    "fastmcp>=2.13.3",
    "google-adk>=1.18.0",
    "langchain>=1.1.3",
    "langchain-google-genai>=4.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.4",
    "mcp>=1.22.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "starlette>=0.50.0",
    "uvicorn>=0.38.0",
]
